Finish layout drawing for 2-zone and 4-zone menu
The two menu are Drawable subclasses of the abstract ZoneMenu class which is responsible for drawing the menu on screen and provide updates for when a selection is made. The protocol for menu selection is yet to be decided but will probably involve communicating an integer index to the parent ExperimentActivity class which will then process and update the next screen accordingly.